Propose rename for future
