+++ vizro-ai/tests/e2e/test_plot.py
